Fix FK_Constraint issue
Previously with CascadeType.PERSIST, whenever a Team was saved
another Participant Entity was saved in the database.

Using mappedBy="currentTeam" makes the Participant the owner of
the relationship and the Team the inverse. Now each time a team
gets saved with a member, a foreign key in Participant get's set
to the teams id