Manually get current user from userService
The current user now always is accessed through the user
service instead of the @AuthenticationPrincipal user: CustomUserDetails.
This is done because these CustomUserDetails seem to be immutably loaded
by the spring security context once and don't change when user data has
changed.
For discussions on this topic see:
http://stackoverflow.com/questions/23072235/reload-userdetails-object-from-database-every-request-in-spring-security
In order to avoid this issue in the future, the CustomUserDetails now don't
implement the User interface anymore, so that one necessarely needs to retreive
the current user by hand, if needed.

This commit fixes related bugs in the test