Model likes as a part of posting
This commit removes the LikeService and LikeRepository,
as postings and likes are modeled as an aggregate now.
Because likes always belong to postings now, the backreference
from like to posting is removed for simplicity. With removing this
backreference, the mapping in the database is transformed from a
foreign key on `postinglike` to a join table.