Model comments as a part of posting
This commit removes the CommentService and CommentRepository,
as postings and comments are modeled as an aggregate now.
Because comments always belong to postings now, the backreference
from comment to posting is removed for simplicity.
With removing this backreference, the mapping in the database
is transformed from a foreign key on `comment` to a join table.