Allow a participant to be part of more than one team
A participant can be part of more than one team, if those teams do not
belong to the same event. This allows participants to partake in
multiple events over time (e.g BreakOut Berlin 2016 & BreakOut München
2017)

This is realized in the model by not only keeping information about the
participants current team `participant.getCurrentTeam()`, but by also
keeping a set of teams that a user has been / or is part of, which can
be received via `participant.getAllTeams()`

This commit adds this behaviours, modifies respecting tests and adds
some tests on the model and the integration level.

It also includes a migration which allows us to add all "current teams"
of participants to the list of all teams the participant has been part
of.

Checks on the service level whether or not a user can create or join a
team have been removed in favor of checking these conditions in the
model.