Add team to posting to allow filtering on team page
As a user can have more than one team (since some time now)
he can have a current team and multiple former teams.
On the team page the postings where shown by selecting all
postings where the creator was someone who had his currentTeam
set to the team that the postings should be displayed for.
This is now fixed by adding a team directly to a posting,
which allows filtering directly by this.